pyproject.toml
[tool.poetry]
name = "exif_photo"
version = "0.1.0"
description = ""
authors = ["Fuminori Hakoishi(ERi19008) <fuminori.hakoishi@erii.co.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
